update enum Const and fix gameObject

+++ src/com/arkanoid/game/GameManager.java
@@ -73,9 +73,9 @@ public class GameManager implements Runnable {
             ball.reverseY();
         }
         if (ball.getBounds().intersects(paddle.getBounds())){
-            int paddleCenterX = paddle.getX() + paddle.getWidth() / 2;
-            int ballCenterX = ball.getX() + ball.getWidth() / 2;
-            int offX = ballCenterX - paddleCenterX;
+            double paddleCenterX = paddle.getX() + paddle.getWidth() / 2;
+            double ballCenterX = ball.getX() + ball.getWidth() / 2;
+            double offX = ballCenterX - paddleCenterX;
             ball.setSpeedX(ball.getMaxSpeed() * offX / (paddle.getWidth() / 2));
             double newSpeedY = Math.sqrt(ball.getMaxSpeed() * ball.getMaxSpeed() - ball.getSpeedX() * ball.getSpeedX());
             ball.setSpeedY(-newSpeedY);
